git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@21813 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/module/main/src/org/geotools/data/jdbc/fidmapper/MaxIncFIDMapper.java
@@ -121,7 +121,7 @@ public class MaxIncFIDMapper extends AbstractFIDMapper {
                     + ") from " + tableName);
 
             if (rs.next()) {
-                int maxFid = rs.getInt(1);
+                long maxFid = rs.getLong(1);
 
                 return String.valueOf(maxFid + 1);
             } else {
